Use freesound.org dataset for benchmarking ...
Unfortunately, since either or neither of the datasets are representative
of the same population, the recommended thresholds for the same tolerated
false-positive rates are different.
This commit intends not to change the behavior, or only marginally, because
QA was happy with the behavior before and we want to spare QA from re-evaluating.
To achieve this, it's the tolerated false-positive rates that we adjusted.
In strict mode, it was raised from 1 to 4%.